+++ CHANGELOG.md
@@ -11,7 +11,8 @@ You can also check the
 
 # Unreleased
 
-Nothing yet.
+- Display medians for power stability in overview page [#479](https://github.com/visualize-admin/electricity-prices-switzerland/issues/479)
+- Display medians for power stability in power stability detail page [#479](https://github.com/visualize-admin/electricity-prices-switzerland/issues/479)
 
 # 2.30.1 - 2025-12-15